requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bandit==1.8.3
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.8.0
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
fastapi==0.115.12
fastapi-health==0.4.0
filelock==3.18.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
license-expression==30.4.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
mypy_extensions==1.1.0
packageurl-python==0.16.0
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.7
pluggy==1.5.0
py-serializable==2.0.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.8
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
stevedore==5.4.1
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
setuptools==65.5.1
psycopg2-binary